// src/app/features/user/models/user.model.ts
export interface User{
    UserId: string;
    Name: string;
    Username: string;
    Birthday: Date;
    PhoneNumber: number;
    Email: string;
    PasswordHash: string;


}

//public Guid UserId { get; set; }
//public string Name { get; set; }
//public string Username { get; set; }
//public DateTime Birthday { get; set; }
//public int PhoneNumber { get; set; }
//public string Email { get; set; }
//public string PasswordHash { get; set; }
//public string Role { get; set; }